# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateBillDto": {
        "properties": {
          "currency": {
            "description": "Three-letter ISO currency code. Defaults to USD.",
            "example": "USD",
            "type": "string"
          },
          "description": {
            "description": "Optional bill description.",
            "example": "Monthly ISP payment",
            "maxLength": 2000,
            "type": "string"
          },
          "dueDate": {
            "description": "Optional due date for this bill.",
            "example": "2026-03-15T18:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "incurredAt": {
            "description": "When the bill was incurred. Defaults to now.",
            "example": "2026-03-05T18:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "paidByUserId": {
            "description": "User id of the member who paid this bill.",
            "example": "f5c6304d-7f58-4f67-bf80-4de6f388b310",
            "type": "string"
          },
          "splits": {
            "description": "Split rows. Amounts must sum exactly to totalAmount.",
            "items": {
              "$ref": "#/components/schemas/CreateBillSplitDto"
            },
            "type": "array"
          },
          "title": {
            "description": "Short bill title.",
            "example": "Internet bill - March",
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "totalAmount": {
            "description": "Total bill amount. Supports up to 2 decimal places.",
            "example": 76.5,
            "type": "number"
          }
        },
        "required": [
          "title",
          "totalAmount",
          "paidByUserId",
          "splits"
        ],
        "type": "object"
      },
      "CreateBillSplitDto": {
        "properties": {
          "amount": {
            "description": "Amount this user owes for the bill. Supports up to 2 decimal places.",
            "example": 25.5,
            "type": "number"
          },
          "userId": {
            "description": "User id for this split row.",
            "example": "f5c6304d-7f58-4f67-bf80-4de6f388b310",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "amount"
        ],
        "type": "object"
      },
...
      "CreatePaymentDto": {
        "properties": {
          "amount": {
            "description": "Payment amount. Supports up to 2 decimal places.",
            "example": 20,
            "type": "number"
          },
          "billId": {
            "description": "Optional bill id this payment is linked to.",
            "example": "3e4f66fd-7eeb-4c2f-a98e-f94376ea22f5",
            "type": "string"
          },
          "currency": {
            "description": "Three-letter ISO currency code. Defaults to USD.",
            "example": "USD",
            "type": "string"
          },
          "idempotencyKey": {
            "description": "Optional idempotency key for safely retried payment requests.",
            "example": "pay-2026-03-05-group1-user1-user2-20",
            "maxLength": 64,
            "type": "string"
          },
          "note": {
            "description": "Optional payment note.",
            "example": "Paid via Zelle",
            "maxLength": 1000,
            "type": "string"
          },
          "paidAt": {
            "description": "When payment was made. Defaults to now.",
            "example": "2026-03-05T18:15:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "payeeUserId": {
            "description": "User id of the member receiving payment.",
            "example": "8e9f7f6e-e3b4-468f-a8f4-c35159b844ab",
            "type": "string"
          },
          "payerUserId": {
            "description": "User id of the member sending payment.",
            "example": "f5c6304d-7f58-4f67-bf80-4de6f388b310",
            "type": "string"
          }
        },
        "required": [
          "payerUserId",
          "payeeUserId",
          "amount"
        ],
        "type": "object"
      },
...
      "UpdateContractDraftDto": {
        "properties": {
          "content": {
            "description": "New draft content for the group contract.",
            "maxLength": 50000,
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Module 1-2 platform and authentication",
    "title": "RoomieManager Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups/{groupId}/balances": {
      "get": {
        "operationId": "FinanceController_getBalances",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns current balances for the group."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute and return net balances (who owes whom) for the group.",
        "tags": [
          "Finance"
        ]
      }
    },
    "/api/v1/groups/{groupId}/bills": {
      "get": {
        "operationId": "FinanceController_listBills",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Maximum number of items per page.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Optional sort field for group bills.",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "enum": [
                "incurredAt",
                "createdAt",
                "totalAmount"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order applied to the selected field.",
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "bills": [
                        {
                          "createdAt": "2026-03-05T14:45:00.000Z",
                          "createdBy": "550e8400-e29b-41d4-a716-446655440001",
                          "currency": "USD",
                          "description": "Monthly ISP payment",
                          "dueDate": null,
                          "groupId": "cm8w9z0abc123def456ghi789",
                          "id": "cm8wc5rr7000hmk6zx4qedcib",
                          "incurredAt": "2026-03-05T14:45:00.000Z",
                          "paidByUserId": "550e8400-e29b-41d4-a716-446655440001",
                          "splitMethod": "CUSTOM",
                          "splits": [
                            {
                              "amount": 25.5,
                              "createdAt": "2026-03-05T14:45:00.000Z",
                              "id": "cm8wc5rr9000imk6z2jmlf4x4",
                              "userId": "550e8400-e29b-41d4-a716-446655440001"
                            }
                          ],
                          "title": "Internet bill - March",
                          "totalAmount": 76.5,
                          "updatedAt": "2026-03-05T14:45:00.000Z"
                        }
                      ],
                      "groupId": "cm8w9z0abc123def456ghi789",
                      "pagination": {
                        "hasNextPage": false,
                        "hasPreviousPage": false,
                        "page": 1,
                        "pageSize": 20,
                        "totalItems": 1,
                        "totalPages": 1
                      }
                    },
                    "meta": {
                      "requestId": "25d65b90-2cb4-451d-a2ce-65f27a223944",
                      "timestamp": "2026-03-05T14:46:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns bill list for the group with pagination metadata."
          },
          "400": {
            "description": "Invalid group ID or invalid pagination/sort query values."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Caller is not an active member of this group."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List bills for a group.",
        "tags": [
          "Finance"
        ]
      },
      "post": {
        "operationId": "FinanceController_createBill",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBillDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Bill created successfully."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a bill and split amounts across group members.",
        "tags": [
          "Finance"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/contract": {
      "get": {
        "operationId": "ContractsController_getContract",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the contract detail."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current contract for a group (draft + latest published).",
        "tags": [
          "Contracts"
        ]
      },
      "put": {
        "operationId": "ContractsController_updateDraft",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns updated contract summary."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the contract draft (admin only).",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/api/v1/groups/{groupId}/contract/publish": {
      "post": {
        "operationId": "ContractsController_publishVersion",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Returns the newly published version."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish the current draft as a new version (admin only).",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/api/v1/groups/{groupId}/contract/versions": {
      "get": {
        "operationId": "ContractsController_listVersions",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Maximum number of items per page.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Optional sort field for contract versions.",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "enum": [
                "version",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order applied to the selected field.",
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "groupId": "cm8w9z0abc123def456ghi789",
                      "pagination": {
                        "hasNextPage": false,
                        "hasPreviousPage": false,
                        "page": 1,
                        "pageSize": 20,
                        "totalItems": 3,
                        "totalPages": 1
                      },
                      "versions": [
                        {
                          "content": "Rent due by 5th. Quiet hours after 10PM.",
                          "createdAt": "2026-03-05T14:44:00.000Z",
                          "id": "cm8wb8l66000fmk6z5nwnk1j5",
                          "publishedBy": "550e8400-e29b-41d4-a716-446655440001",
                          "version": 3
                        }
                      ]
                    },
                    "meta": {
                      "requestId": "8f9c8a41-26b8-4e53-a35e-64f4db76d79e",
                      "timestamp": "2026-03-05T14:45:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns version history with pagination metadata."
          },
          "400": {
            "description": "Invalid group ID or invalid pagination/sort query values."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Caller is not an active member of this group."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all published contract versions for a group.",
        "tags": [
          "Contracts"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/payments": {
      "post": {
        "operationId": "FinanceController_createPayment",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Payment recorded successfully."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a payment between group members.",
        "tags": [
          "Finance"
        ]
      }
    },
...
  },
  "servers": [],
...
}